third_party/blink/web_tests/external/wpt/css/css-conditional/container-queries/multiple-style-containers-comma-separated-queries-expected.txt

This is a testharness.js-based test.
[FAIL] Should match the named outer container for --foo:bar
  assert_equals: expected "yes" but got "no"
[FAIL] Match the #combined container for --foo:qux which is also a size container
  assert_equals: expected "yes" but got "no"
Harness: the test ran to completion.
